freebsd-src-regression-suite-run: Create toolchain symlinks
Some tests (at least dtrace and libatf) expect standard C toolchain
executables in the PATH.  We install llvm anyway, so let's provide them.

This could and perhaps should be done when the image is generated from
the mtree manifest.  But, at that point the llvm package isn't installed
yet; it's installed during first boot.  So, we'd be adding dangling
symlinks, and I'm not sure yet if makefs supports that.  For now, it's
easy to do it during VM boot.