freebsd-src-build: Split mtree files when installing
This is a step towards having a tighter edit-build-test loop: when an
installworld is done, its mtree and stagedir are left alone, and
subsequent buildkernel+installkernel builds can be combined with that
result.  The upshot of this is that freebsd-src-build:make_targets can
be set to "buildkernel installkernel" and we still get a useful VM image
from the result.  This is much faster than is possible otherwise, though
there's still a lot of room for improvement since we still need to
rebuild the filesystem and disk images each run.