syzkaller-freebsd-fuzz: Put the kernel config in /tmp
Add a KERNCONFDIR parameter to freebsd-src-build as well.  This way we
don't pollute the src tree with an extra uncommitted file.

Putting in /tmp is not ideal, I'd rather have it under the
syzkaller-freebsd-fuzz workdir, but bricoler doesn't have a well-defined
per-task path variable that we can expand to an absolute path at task
definition time.