Add a new implementation of bricoler
This time, in python.  The extensive standard library and affordances
related to types make this much more pleasant than lua, at the cost of
speed, but that's the right tradeoff.

The underlying implementation is quite similar.  There are some changes
to the command-line syntax:
1) There are no more verbs, just run `bricoler <task> <flags> <params>`.
   Some flags can change behaviour, e.g., one can have

   $ bricoler freebsd-src-build -s

   to show the parameter bindings.
2) Parameter syntax is a bit shorter, instead of
   `--param <task>:<param>=<val>`, use `--<task>/<param>=<val>`.

This is far from being complete, I am still porting over much of the
meat from the lua implementation.  But most of the core is there.  I
also added a new feature: one can now define aliases for useful
commands, so I can, for example, make "bricoler foo" run a task with a
whole set of parameters.  Aliases are stored in bricoler.json at the
root of the workdir.  To define an alias, run:

   $ bricoler <task> <params> -a <alias>

Now you can run `bricoler <alias>` instead of
`bricoler <task> <params>`.

This attempt is far enough along that I'll probably finish it and
replace the lua version, so let's start checkpointing.